Use 64 bit floats.

+++ vm/defs.hpp
@@ -10,7 +10,7 @@ namespace vm {
 typedef unsigned short field_num;
 typedef uint32_t uint_val;
 typedef int32_t int_val;
-typedef float float_val;
+typedef double float_val;
 typedef uint64_t ptr_val;
 typedef uint32_t node_val;
 typedef uint32_t worker_val;
